fix integration test installation
